requirements.txt
Flask
Flask-SQLAlchemy
Flask-Admin
Flask-Babel
mysql-connector-python
pandas
numpy
requests
beautifulsoup4
snownlp
geopy
pymysql
pyecharts
python-dotenv